Move backup related screens in features/backup folder
