+++ apps/mobile-wallet/src/features/backup/backupScreens/VerificationSuccessScreen.tsx
