+++ apps/mobile-wallet/src/features/backup/backupScreens/VerifyMnemonicScreen.tsx
